Rename folder "data" to "manageData".
