+++ src/lib/database/current/manageData.ts
@@ -3,12 +3,12 @@ import { CoffeeBeans } from "$lib/domain/entities/CoffeeBeans";
 import { Recipe, RecipeSubmit } from "$lib/domain/entities/Recipe";
 import type { Count } from "$lib/types/Count";
 import { COFFEEBEANS_STORE_NAME, RECIPES_STORE_NAME, openEntitiesDB } from "./core/indexedDbCore";
-import { DemoCoffeeBeans } from "./data/demo/demoCoffeeBeans";
-import { generateDemoRecipes } from "./data/demo/demoRecipes";
-import { matchUniqueCoffeeBeansToAdd, matchUniqueRecipesToAdd } from "./data/import/match/arrays";
-import { parseCoffeeBeansArray, parseRecipesArray } from "./data/import/parse/arrays";
-import { parseDbVersion } from "./data/import/parse/primitives";
-import { vacuumSoftDeletedCoffeeBeans, vacuumSoftDeletedRecipes } from "./data/vacuum";
+import { DemoCoffeeBeans } from "./manageData/demo/demoCoffeeBeans";
+import { generateDemoRecipes } from "./manageData/demo/demoRecipes";
+import { matchUniqueCoffeeBeansToAdd, matchUniqueRecipesToAdd } from "./manageData/import/match/arrays";
+import { parseCoffeeBeansArray, parseRecipesArray } from "./manageData/import/parse/arrays";
+import { parseDbVersion } from "./manageData/import/parse/primitives";
+import { vacuumSoftDeletedCoffeeBeans, vacuumSoftDeletedRecipes } from "./manageData/vacuum";
 import { CoffeeBeansDB, CoffeeBeansDBSubmit, type ICoffeeBeansDB } from "./types/CoffeeBeansDB";
 import type { EntitiesDB } from "./types/EntitiesDB";
 import type { ExportJSON } from "./types/ExportJSON";
